+++ ironweb.d.ts
@@ -203,10 +203,6 @@ export interface BlindSearchIndex {
  * SDK Namespaces
  */
 export interface User {
-    /**
-     * Deprecated, use deleteDevice() with no argument to deauthorize the current device instead.
-     */
-    deauthorizeDevice(): Promise<{transformKeyDeleted: boolean}>;
     deleteDevice(deviceId?: number): Promise<number>;
     deleteDeviceByPublicSigningKey(publicSigningKey: Base64String): Promise<number>;
     listDevices(): Promise<UserDeviceListResponse>;